Update Fix Detail

+++ app/src/main/java/dev/byto/moly/domain/model/ReviewList.kt
@@ -31,5 +31,5 @@ data class Author(
     val name: String?,
     val username: String?,
     val avatarPath: String?,
-    val rating: Int?
+    val rating: Double?
 ) : Parcelable
